+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Allow to set expiration delta for `TransactionRequest` (#553).
 * Implemented `GetAccountProof` endpoint (#556)
+* Added WASM consumable notes API + improved note models (#561).
 * [BREAKING] Refactored `OutputNoteRecord` to use states and transitions for updates (#551).
 * Added better error handling for WASM sync state (#558).
 * Added WASM Input note tests + updated input note models (#554)